requirements.txt
chromadb
python-dotenv
openai
PyMuPDF
requests